add functionalities refactor & readme according new project && git push
